IP2Proxy

Description

This Haskell package allows users to query an IP address to determine if it was being used as open proxy, web proxy, VPN anonymizer and TOR exits.

IP2Proxy LITE BIN databases are available for free at http://lite.ip2location.com/

Documentation

data Meta

Contains the BIN database file metadata.

data IP2ProxyRecord

Contains proxy results.

Constructors

IP2ProxyRecord

Fields

country_short :: String

Country code

country_long :: String

Country name

region :: String

Region name

city :: String

City name

isp :: String

ISP name

proxy_type :: String

Proxy type

is_proxy :: Int

Is proxy

getModuleVersion :: String

The getModuleVersion function returns a string containing the module version.

getPackageVersion :: Meta -> String

The getPackageVersion function returns a string containing the package version. It takes 1 argument; the metadata from open function (Meta record).

getDatabaseVersion :: Meta -> String

The getDatabaseVersion function returns a string containing the database version. It takes 1 argument; the metadata from open function (Meta record).

open :: String -> IO Meta

The open function returns the Meta record containing metadata from the BIN database file. It takes one argument, of type String, which is the path to the BIN database file.

getAll :: String -> Meta -> String -> IO IP2ProxyRecord

The getAll function returns an IP2ProxyRecord containing proxy data for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

getCountryShort :: String -> Meta -> String -> IO String

The getCountryShort function returns the country code for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

getCountryLong :: String -> Meta -> String -> IO String

The getCountryLong function returns the country name for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

getRegion :: String -> Meta -> String -> IO String

The getRegion function returns the region name for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

getCity :: String -> Meta -> String -> IO String

The getCity function returns the city name for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

getISP :: String -> Meta -> String -> IO String

The getISP function returns the ISP name for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

getProxyType :: String -> Meta -> String -> IO String

The getProxyType function returns the proxy type for an IP address. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).

isProxy :: String -> Meta -> String -> IO String

The isProxy function returns 0 if IP is not a proxy, 1 if is a proxy and not data center IP, 2 if is a proxy and is a data center IP, -1 if error. It takes 3 arguments; the BIN database file path (String), the metadata from open function (Meta record) & either IPv4 or IPv6 address (String).